requirements.txt
acme==1.1.0
asgiref==3.8.1
attrs==24.2.0
Automat==0.8.0
backports.zoneinfo==0.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.4
Brotli==1.1.0
certbot==0.40.0
certbot-apache==0.39.0
certbot-nginx==0.40.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
Click==7.0
cloud-init==24.3.1
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cssselect2==0.7.0
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.16
django-cron==0.6.0
docopt==0.6.2
entrypoints==0.3
et-xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.55.3
future==0.18.2
h11==0.14.0
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==3.10
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
mock==3.0.5
more-itertools==4.2.0
mysql-connector-python==9.0.0
mysqlclient==2.2.6
netifaces==0.10.4
num2words==0.5.14
numpy==1.24.4
oauthlib==3.1.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
parsedatetime==2.4
pbr==5.4.5
pexpect==4.6.0
pillow==10.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.26.1
pycparser==2.22
pydyf==0.11.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.3.0
pyodbc==5.2.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyphen==0.16.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-augeas==0.5.0
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==5.3.1
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.27.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.7.2
soupsieve==2.6
sqlparse==0.5.2
ssh-import-id==5.10
systemd-python==234
tinycss2==1.4.0
trio==0.27.0
trio-websocket==0.11.1
Twisted==18.9.0
typing-extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.3
weasyprint==61.2
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
zopfli==0.2.3.post1
